Code/server/requirements.txt
audioread
huggingface-hub
importlib_metadata
librosa
openai-whisper
safetensors
scikit-learn
scipy
sentencepiece
soundfile
tiktoken
tokenizers
transformers



